Rename effects and add animation curve
